uploadCheckpoint uploads on every checkpoint epoch
